+++ android/app/src/main/AndroidManifest.xml
@@ -51,7 +51,7 @@
           <action android:name="android.intent.action.VIEW" />
           <category android:name="android.intent.category.DEFAULT" />
           <category android:name="android.intent.category.BROWSABLE" />
-          <data android:scheme="turtlecoin" android:host="" /> // A
+          <data android:scheme="turtlecoin"/>
         </intent-filter>
       </activity>
       <activity android:name="com.facebook.react.devsupport.DevSettingsActivity" />
